removed utulities file instead of helpers file, to allow for better building proicess
