+++ CMakeLists.txt
@@ -82,7 +82,7 @@ list(INSERT CMAKE_MODULE_PATH 0 ${CMAKE_SOURCE_DIR}/cmake/Modules)
 # Set the version information here
 set(VERSION_MAJOR 0)
 set(VERSION_API   3)
-set(VERSION_ABI   2)
+set(VERSION_ABI   3)
 set(VERSION_PATCH git)
 
 cmake_policy(SET CMP0011 NEW)
