+++ lib/add_crc_impl.h
@@ -2,7 +2,7 @@
 #define INCLUDED_LORA_ADD_CRC_IMPL_H
 
 #include <lora_sdr/add_crc.h>
-#include <lora_sdr/utilities.h>
+#include "helpers.h"
 namespace gr {
 namespace lora_sdr {
