+++ lib/data_source_impl.h
@@ -2,7 +2,7 @@
 #define INCLUDED_LORA_SDR_DATA_SOURCE_IMPL_H
 
 #include <lora_sdr/data_source.h>
-#include <lora_sdr/utilities.h>
+#include "helpers.h"
 
 namespace gr {
 namespace lora_sdr {
