+++ lib/hamming_dec_impl.cc
@@ -5,7 +5,7 @@
 
 #include "hamming_dec_impl.h"
 #include <gnuradio/io_signature.h>
-#include <lora_sdr/utilities.h>
+#include "helpers.h"
 
 namespace gr {
 namespace lora_sdr {
