+++ lib/hamming_enc_impl.cc
@@ -4,7 +4,7 @@
 
 #include "hamming_enc_impl.h"
 #include <gnuradio/io_signature.h>
-#include <lora_sdr/utilities.h>
+#include "helpers.h"
 
 namespace gr {
 namespace lora_sdr {
