+++ lib/modulate_impl.h
@@ -6,7 +6,7 @@
 #include <iostream>
 #include <lora_sdr/modulate.h>
 
-#include <lora_sdr/utilities.h>
+#include "helpers.h"
 
 namespace gr {
 namespace lora_sdr {
